% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{addturn}
\alias{addturn}
\alias{closest2end}
\alias{find.LOS.nodes}
\alias{dist2nodes}
\alias{find.unusables}
\title{Helper functions for routeKFS}
\usage{
addturn(currentnode)

closest2end(startpt, endpt, nextnodes, nodepts)

find.LOS.nodes(node, nodepts, toprint = TRUE)

dist2nodes(pt1, pt2)

find.unusables(currentID, currentnode, endpt, nextnodes, nodepts)
}
\description{
Helper functions for routeKFS
}
\section{Functions}{
\itemize{
\item \code{closest2end}: Determine node closest to end destination

\item \code{find.LOS.nodes}: Determine which nodes are line-of-sight of a point

\item \code{dist2nodes}: Calculated distance between two nodes

\item \code{find.unusables}: Eliminate potential nodes based on their distance from the current node
}}

