% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bangarang.R
\docType{package}
\name{bangarang}
\alias{bangarang}
\alias{bangarang-package}
\title{bangarang: A package of R tools for data from the Kitimat Fjord System}
\description{
bangarang is a collection of functions useful
in the Kitimat Fjord System, British Columbia
(Gitga'at First Nation marine territory),
though some functions are applicable to the
northeast Pacific coast in general.
Developed for the Bangarang Project
but generalized for use by the
Gitga'at Oceanographic Initiative,
NCCS, DFOs and others.
}
\section{Primary functions}{

The primary functions in this package are:
\describe{
\item{\code{plotKFS}}{Mapping tools}
\item{\code{whalemap}}{Locating a whale in confined channels based on survey data}
\item{\code{routeKFS}}{Travel distance between two points in the fjord system (around islands)}
}
}

\section{Secondary functions}{

There are some secondary functions used by these primaries,
and they may be useful in and of themselves:
\describe{
\item{\code{solve2lines}}{Determine the intersection point of two lines}
\item{\code{LOStest}}{Determine whether two points are within line-of-sight
of each other or whether the shoreline is in the way}
}
}

\section{Background functions}{

These functions also have small helper functions that
probably don't have a use otherwise or on their own:
\describe{
\item{\code{calcOD}}{in \code{routeKFS}}
\item{\code{addturn}}{in \code{routeKFS}}
\item{\code{closest2end}}{in \code{routeKFS}}
\item{\code{find.LOS.nodes}}{in \code{routeKFS}}
\item{\code{dist2nodes}}{in \code{routeKFS}}
\item{\code{find.unusables}}{in \code{routeKFS}}
\item{\code{oceandistance}}{in \code{routeKFS}}
\item{\code{plotOD}}{in \code{routeKFS}}
}
}
\author{
Eric Keen, Scripps Institution of Oceanography, \email{ekeen@ucsd.edu}
}

