% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc-blocks.R
\docType{data}
\name{blocks}
\alias{blocks}
\title{Boundaries for geographic block schemes in Bangarang analyses}
\format{
A data frame with 54 rows and 11 columns:
\describe{
  \item{ID}{Sequential identifier}
  \item{name}{Descriptive name of the block location}
  \item{sq.km}{Square area of the water within the rectangle.}
  ...
}
}
\source{
Written and compiled by Eric Keen
}
\usage{
blocks
}
\description{
A dataset containing the names, locations, and plotting details of oceanographic stations and transet
turning points of the Bangarang study plan.
}
\keyword{datasets}
