% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapper.R
\name{bootstrapper}
\alias{bootstrapper}
\title{Create a bootstrap of segments and their sightings}
\usage{
bootstrapper(segments, sightings)
}
\arguments{
\item{segments}{`data.frame` of segment summary data (each row is a segment),
with a columns named `Sample.Label`.}

\item{sightings}{`data.frame` of sightings (each row is a sighting),
with a column named `Sample.Label`.}
}
\value{
A `list` with two slots: a bootstrap version of `segments` and a bootstrap version of `sightings`. 
Each `data.frame` has a revised `Sample.Label` column and a new column, `old_Sample`, which records the original sample label.
}
\description{
Create a bootstrap of segments and their sightings
}
