% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapper_summary.R
\name{bootstrapper_summary}
\alias{bootstrapper_summary}
\title{Bootstrap summary}
\usage{
bootstrapper_summary(bootstraps)
}
\arguments{
\item{bootstraps}{Dataframe of bootstrap results}
}
\value{
Dataframe table
}
\description{
Bootstrap summary
}
