% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapper_wrapper.R
\name{bootstrapper_wrapper}
\alias{bootstrapper_wrapper}
\title{Wrapper for bootstrap routine for density estimation for each circuit}
\usage{
bootstrapper_wrapper(
  dsm_formula,
  segments,
  observation_data,
  B = 1000,
  output_filename = "bootstraps"
)
}
\arguments{
\item{dsm_formula}{String}

\item{segments}{Segment data}

\item{observation_data}{Formatted observation (sightings) data}

\item{B}{Number of iterations}

\item{output_filename}{If you want the result to be saved to a data file in 
your working directory, specify filename here. File type will be `.rds`.}
}
\value{
Dataframe of bootstrap results.
}
\description{
Wrapper for bootstrap routine for density estimation for each circuit
}
