% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_flight_flux.R
\name{calculate_flight_flux}
\alias{calculate_flight_flux}
\title{Calculate flight flux correction factor}
\usage{
calculate_flight_flux(sightings, flight_groups)
}
\arguments{
\item{sightings}{Seabird sightings.}

\item{flight_groups}{Table of wind-contingent speed data for seabird groups (see `data(flight_groups)`).}
}
\value{
Input `data.frame` with new columns.
}
\description{
Calculate flight flux correction factor
}
