% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc-channels.R
\docType{data}
\name{channels}
\alias{channels}
\title{Polygons of coordinates delineating 8 geographic channels}
\format{
A data frame with 4811 rows and 5 columns, formatted for seamless mapping with the PBSmapping package.
}
\source{
Written and compiled by Eric Keen
}
\usage{
channels
}
\description{
A dataframe of coordinates for mapping the 8 geographic blocks used to partition the Kitimat 
Fjord System in Bangarang analyses.
}
\keyword{datasets}
