% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_detection_function.R
\name{create_dummy_detection_function}
\alias{create_dummy_detection_function}
\title{Create dummy detection function}
\usage{
create_dummy_detection_function(observation_data)
}
\arguments{
\item{observation_data}{Sightings, formatted}
}
\value{
Detection function object from package `dsm`
}
\description{
Create dummy detection function
}
