% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsm_compare.R
\name{dsm_compare}
\alias{dsm_compare}
\title{Fit a DSM with permutations of covariates then rank results}
\usage{
dsm_compare(
  segments,
  observation_data,
  dummy_ds,
  candidate_splines,
  base_formula = "density.est ~ s(x,y)"
)
}
\arguments{
\item{segments}{Segments}

\item{observation_data}{Formatted observation data}

\item{dummy_ds}{Dummy detection function}

\item{candidate_splines}{Character vector of candidate splines}

\item{base_formula}{Base formula}
}
\value{
A list with a slot for ranked summary and a slot for the model objects themselves
}
\description{
Fit a DSM with permutations of covariates then rank results
}
