% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc-effort.R
\docType{data}
\name{effort}
\alias{effort}
\title{Bangarang survey effort}
\format{
An object of class \code{data.frame} with 25526 rows and 27 columns.
}
\source{
Written and compiled by Eric Keen Ezell.
}
\usage{
effort
}
\description{
A dataset containing position updates for all survey effort aboard the Bangarang, 2013 - 2015.
}
\keyword{datasets}
