% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_observation_data.R
\name{format_observation_data}
\alias{format_observation_data}
\title{Format sightings into observation data for `dsm`}
\usage{
format_observation_data(sightings)
}
\arguments{
\item{sightings}{Dataframe of Bangarang sightings.}
}
\value{
Observation data formatted.
}
\description{
Format sightings into observation data for `dsm`
}
