% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_transects.R
\name{generate_transects}
\alias{generate_transects}
\title{Generate a transect plan}
\usage{
generate_transects(
  write_to = ".",
  show_plot = TRUE,
  save_to_file = TRUE,
  verbose = TRUE,
  my_crs = 4326
)
}
\arguments{
\item{write_to}{The directory to which files should be saved; default is your working directory.}

\item{show_plot}{Print plot of transect plan?}

\item{save_to_file}{Save transect plan to file?}

\item{verbose}{Print status updates to the Console?}

\item{my_crs}{The projection you wish to use when writing the layout to file. Default is WS84, whicih is suitable for British Columbia, Canada.}
}
\value{
Unless options are turned off, this function will display a map of the newly generated transect layout, then save that transect layout to a 
map file (PDF), a spreadsheet (CSV), and to a GPX file, which can then be loaded onto a boat chartplotter of handheld GPS in order to guide observers along the transects.
}
\description{
This function generates a new random layout of transect lines within the Kitimat Fjord System study area
using the study design generated with the `R` package `dssd` and saved in the `bangarang` dataset `study_design`, 
which is meant to guide long-term line-transect distance sampling surveys by Gitga'at collaborators for the years 2024 and beyond.
}
