% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_kfs.R
\name{gg_kfs}
\alias{gg_kfs}
\title{`ggplot()` of the Kitimat Fjord System}
\usage{
gg_kfs(lat_range = c(52.8, 53.55), lon_range = c(-129.68, -128.85))
}
\arguments{
\item{lat_range}{Two-element vector of latitudinal range}

\item{long_range}{Two-element vector of longitudinal range}
}
\value{
Returns a `ggplot()` map object.
}
\description{
`ggplot()` of the Kitimat Fjord System
}
