% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfs_chloropleth.R
\name{gg_kfs_grid}
\alias{gg_kfs_grid}
\title{Plot a chloropleth surface in the KFS}
\usage{
gg_kfs_grid(
  values,
  kfs_grid = NULL,
  point_size = 2.35,
  color_map = "viridis",
  scale_breaks = NULL,
  scale_limits = NULL,
  pseudo_log = FALSE,
  map_title = NULL,
  legend_title = "Values",
  lat_range = c(52.8, 53.55),
  lon_range = c(-129.68, -128.85),
  land_fill = I("#d5dcde"),
  land_line = "grey30",
  land_stroke = 0.25,
  land_alpha = 1,
  water_fill = "lightblue",
  water_alpha = 0.1
)
}
\arguments{
\item{values}{The vector of values to plot. Length must be same number of rows as `kfs_grid` input.}

\item{kfs_grid}{`data.frame` of grid coordinates, with `x` and `y` columns. If `NULL`, the built-in `grid` dataset from the `bangarang` package will be used.}

\item{point_size}{Size of each grid point. May need to be adjusted based on output size of your plot.}

\item{color_map}{The color palette to use from the `viridis` package.}

\item{scale_breaks}{Vector of values to note in the color scale legend.}

\item{scale_limits}{Two-element vector of minimum and maximum values to plot.}

\item{pseudo_log}{Boolean; if `TRUE`, the color scale will be log-transformed to better-show minute variations at small values. Zero will still be plotted.}

\item{map_title}{Optional.}

\item{legend_title}{Optional.}

\item{lat_range}{Two-element vector of latitudinal range}

\item{land_fill}{Land fill color}

\item{land_line}{Land border color}

\item{land_stroke}{Land border thickness}

\item{land_alpha}{Land fill alpha}

\item{water_fill}{Water fill color}

\item{water_alpha}{Water fill alpha}

\item{long_range}{Two-element vector of longitudinal range}
}
\value{
Plot.
}
\description{
Plot a chloropleth surface in the KFS
}
