% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_size_draw.R
\name{group_size_draw}
\alias{group_size_draw}
\title{Random draw from truncated log-normal distribution based on min, best, max group size estimte}
\usage{
group_size_draw(sightings)
}
\arguments{
\item{sightings}{Dataframe of sightings}
}
\value{
Modified sightings dataframe, with the `best` column updated with the random draws.
}
\description{
Random draw from truncated log-normal distribution based on min, best, max group size estimte
}
