% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_block.R
\name{in_block}
\alias{in_block}
\title{Assign coordinates to KFS block}
\usage{
in_block(x, y, toplot = FALSE)
}
\arguments{
\item{x}{Longitude, decimal degrees.}

\item{y}{Latitude, decimal degrees.}

\item{toplot}{Boolean with default `FALSE`.}
}
\value{

}
\description{
Assign coordinates to KFS block
}
