% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_kfs.R
\name{in_kfs}
\alias{in_kfs}
\title{Are coordinates within the KFS?}
\usage{
in_kfs(xy, toplot = FALSE)
}
\arguments{
\item{xy}{Dataframe with coordinates, decimal degrees, in columns 'x' and 'y'}

\item{toplot}{Boolean with default `FALSE`.}
}
\value{
The `xy` dataframe, filtered only to coordinates that are in the KFS.
}
\description{
Are coordinates within the KFS?
}
