% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc-kfs_blocks_bbox.R
\docType{data}
\name{kfs_blocks_bbox}
\alias{kfs_blocks_bbox}
\title{Bounding box for 24 "Bangarang" geostrata (or "blocks")}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 26 rows and 5 columns.
}
\usage{
kfs_blocks_bbox
}
\description{
Bounding box for 24 "Bangarang" geostrata (or "blocks")
}
\keyword{datasets}
