% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc-kfs_boundary.R
\docType{data}
\name{kfs_boundary}
\alias{kfs_boundary}
\title{Polygon representing the boundary of the Kitimat Fjord System}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 1 rows and 1 columns.
}
\usage{
kfs_boundary
}
\description{
Polygon representing the boundary of the Kitimat Fjord System
}
\keyword{datasets}
