% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc-kfs_land.R
\docType{data}
\name{kfs_land}
\alias{kfs_land}
\title{Polygons representing the shoreline of islands within the Kitimat Fjord System polygon}
\format{
An object of class \code{SpatialPolygons} of length 1.
}
\usage{
kfs_land
}
\description{
Polygons representing the shoreline of islands within the Kitimat Fjord System polygon
}
\keyword{datasets}
