% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc-kfs_seafloor.R
\docType{data}
\name{kfs_seafloor}
\alias{kfs_seafloor}
\title{Dataframe of seafloor data for the Kitimat Fjord System}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 375178 rows and 3 columns.
}
\usage{
kfs_seafloor
}
\description{
Dataframe of seafloor data for the Kitimat Fjord System
}
\keyword{datasets}
