% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc-nepac_shore_sf.R
\docType{data}
\name{nepac_shore_sf}
\alias{nepac_shore_sf}
\title{SF object of NEPAC shoreline from PBSmapping}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 2 columns.
}
\usage{
nepac_shore_sf
}
\description{
SF object of NEPAC shoreline from PBSmapping
}
\keyword{datasets}
