% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc-ocean.R
\docType{data}
\name{ocean}
\alias{ocean}
\title{Bangarang survey effort}
\format{
An object of class \code{data.frame} with 1488101 rows and 7 columns.
}
\source{
Written and compiled by Eric Keen Ezell.
}
\usage{
ocean
}
\description{
A dataset containing oceanographic variables throughout an interpolated grid of the 
Bangarang study area for each circuit of each year.
}
\keyword{datasets}
