% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oceandistance.R
\name{ocean.route}
\alias{ocean.route}
\title{Primary functions for routeKFS}
\usage{
ocean.route(pts.ocean, buffer = 0.01, LOSplots = FALSE,
  show.progress = TRUE)
}
\description{
Primary functions for routeKFS
}

