% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotkfs.R
\name{plotKFS}
\alias{plotKFS}
\title{Maps for the Kitimat Fjord System}
\usage{
plotKFS(
  area = "Study Area",
  X = c(-129.68, -128.85),
  Y = c(52.8, 53.55),
  mar = c(4.2, 3, 0.2, 0.2),
  mai = c(0.75, 0.75, 0.1, 0.1),
  land.col = "light gray",
  shore.border = "dark gray",
  shore.lwd = 1,
  tankers = FALSE,
  tankers.width = 5,
  tankers.col = "tan1",
  places = FALSE,
  places.scale = 1,
  stat = FALSE,
  stat.col = "black",
  stat.scale = 1,
  tracks = FALSE,
  track.col = "black",
  blocking = NULL,
  block.col = "black",
  block.lwd = 1,
  block.lab.scale = FALSE
)
}
\arguments{
\item{area}{Map range. Default is "Study Area" of the Bangarang project. Other range options:
\itemize{
  \item "Caamano"
  \item "Estevan"
  \item "Campania"
  \item "Squally"
  \item "S. Squally" (Dougan to McDonald)
  \item "N. Squally" (McDonald to Union)
  \item "Whale"
  \item "Wright"
  \item "McKay" (includes S. Ursula up to Bishop)
  \item "Verney" (includes N. Ursula down to Bishop)
  \item "Gil Basin" (the channels surrounding Gil Island)
  \item "Gribbell Island" (the channels surround Gribbell)
  \item "Other" (custom range, entered using \code{X} and \code{Y} inputs)
}}

\item{X}{Custom longitude range (e.g., \code{c(-129.68,-128.85)}). Only applies when "Other" is entered as area.}

\item{Y}{Custom latitude range (e.g., \code{c(52.9,53.3)}). Only applies when "Other" is entered as area.}

\item{mar}{Plot margins, passed to \code{\link{par}}.}

\item{mai}{Margin size specific in inches, passted to \code{\link{par}}.}

\item{land.col}{Land color.}

\item{shore.border}{Color of shoreline border.}

\item{shore.lwd}{Shoreline border thickness.}

\item{tankers}{If \code{TRUE} (not default), proposed tanker route is displayed.}

\item{tankers.width}{Tanker track line width.}

\item{tankers.col}{Tanker track color}

\item{places}{If \code{TRUE} (not default), research locations are labeled: Hartley Bay, Whale Point, GW Camp, and the Wall.}

\item{places.scale}{Text size of place labels.}

\item{stat}{If \code{TRUE} (not default), shows grid of 24 Bangarang stations.}

\item{stat.col}{Station color.}

\item{stat.scale}{Size of station dots.}

\item{tracks}{If \code{TRUE} (not default), shows Bangarang survey tracklines.}

\item{track.col}{Trackline color.}

\item{blocking}{Controls which block stratification, if any, to show. Default is nothing. 
If \code{"mini"}, shows a 54-block scheme. 
\code{"minirect"} shows the rectangles used to partition the study area into those blocks. 
\code{"sub"} shows a 24-block scheme, the scheme used in the HW-Wave study.
\code{"channel"} shows an 8-block scheme.
\code{"province"} combines those 8 channels into 4 large provinces: offshore, central outer, central inner, and inland.}

\item{block.col}{Color of blocking borders.}

\item{block.lwd}{Blocking border width.}

\item{block.lab.scale}{Default is \code{FALSE}. If a number, it prints labels within the blocks (only for "mini" and "minirect"), scaled to the number given.}
}
\value{
Generates map.
}
\description{
Mapping functions in the Kitimat Fjord System tailored to research needs
of the Bangarang, Gitga'at, and NCCS teams.
}
\examples{
##############################################
 # Simple plot of study area
 plotKFS()
##############################################
# Zoom in of Squally Channel
plotKFS(area="Squally")
##############################################
# Self-defined map range: Kitimat Fjord System with tanker route and study tracklines
plotKFS(area="Other",
       X=c(-129.9,-127.9),
       Y=c(52.5,54.05),
       tankers=TRUE,
       tankers.col=adjustcolor("firebrick",alpha.f=.4),
       tracks=TRUE)
##############################################
# Research headquarters
plotKFS(places=TRUE,places.scale=1.5)
##############################################
# 2015 Bangarang sampling plan
plotKFS(stat=TRUE,stat.scale=1.5,tracks=TRUE)
##############################################
# Examples of study area blocking for various analyses
plotKFS(blocking="minirect",block.lab.scale=1.5) # 54 mini blocks -- never used in analysis
plotKFS(blocking="mini") # Shoreline within those 54 miniblocks
plotKFS(blocking="sub") # Each waterway split into 3 subblocks
plotKFS(blocking="channel") # The 8 main waterways of Bangarang analyses
plotKFS(blocking="province") # Four main provinces
}
\author{
Eric Keen, Scripps Institution of Oceanography, \email{ekeen@ucsd.edu}
}
