% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_summary.R
\name{point_summary}
\alias{point_summary}
\title{Point estimate summary}
\usage{
point_summary(point_estimate)
}
\arguments{
\item{point_estimate}{Dataframe of point estimates the same length as the number of rows in `grid`}
}
\value{
Summary table
}
\description{
Point estimate summary
}
