% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_circuits.R
\name{predict_circuits}
\alias{predict_circuits}
\title{Predict DSM densities for each circuit of effort}
\usage{
predict_circuits(
  dsm_keep,
  segments,
  grid,
  circuits = NULL,
  cex_scale = 1,
  toplot = FALSE
)
}
\arguments{
\item{dsm_keep}{DSM model object to use.}

\item{segments}{Dataframe of segment summary data.}

\item{grid}{Grid dataframe: see `data(grid)`}

\item{circuits}{Dataframe with start/end timestamps for each circuit. See `data(circuit)` for reference.}

\item{cex_scale}{Scale points on DSM maps to be more legible using a number (1 = no scaling).}

\item{toplot}{`TRUE` or `FALSE`?}
}
\value{
A dataframe with the grid replicated for each circuit, with an additional column, 
`d`, containing the density estimate for each grid-cell in each circuit.
}
\description{
Predict DSM densities for each circuit of effort
}
