% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dsm_kfs.R
\name{predict_dsm_kfs}
\alias{predict_dsm_kfs}
\title{Predict density surface for segments using `dsm` model}
\usage{
predict_dsm_kfs(dsm_keep, segments, grid, cex_scale = 1, toplot = TRUE)
}
\arguments{
\item{dsm_keep}{Model object to use for prediction}

\item{segments}{Segment data to predict on}

\item{grid}{Grid dataframe (see `data(grid)`)}

\item{cex_scale}{Scaling factor to scale the points in the plot.}

\item{toplot}{True or False}
}
\value{
Vector of density estimates to match the rows in the `grid` input
}
\description{
Predict density surface for segments using `dsm` model
}
