% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-doc-provinces.R
\docType{data}
\name{provinces}
\alias{provinces}
\title{Polygons of coordinates delineating 4 geographic provinces}
\format{A data frame with 4458 rows and 5 columns, formatted for seamless mapping with the PBSmapping package.}
\source{
Written and compiled by Eric Keen
}
\usage{
provinces
}
\description{
A dataframe of coordinates for mapping the 4 geographic regions ("provinces") used to partition the Kitimat
Fjord System in Bangarang analyses.
}
\keyword{datasets}

