% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc-salmon.R
\docType{data}
\name{salmon}
\alias{salmon}
\title{Bangarang salmon jump detections}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1033 rows and 14 columns.
}
\source{
Written and compiled by Eric Keen Ezell.
}
\usage{
salmon
}
\description{
All detections of salmon jumps for all survey effort aboard the Bangarang, 2013 - 2015.
}
\keyword{datasets}
