% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc-seabirds.R
\docType{data}
\name{seabirds}
\alias{seabirds}
\title{Bangarang seabird sightings}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3668 rows and 30 columns.
}
\source{
Written and compiled by Eric Keen Ezell.
}
\usage{
seabirds
}
\description{
All seabird sightings for all survey effort aboard the Bangarang, 2013 - 2015.
}
\keyword{datasets}
