% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-doc-shiplane.R
\docType{data}
\name{shiplane}
\alias{shiplane}
\title{Coordinates of proposed shipping lane in Kitimat Fjord System}
\format{A data frame with 137 rows and 4 columns, formatted for seamless mapping with the PBSmapping package.}
\source{
Written and compiled by Eric Keen
}
\usage{
shiplane
}
\description{
A dataframe of coordinates that trace a proposed shipping lane in the Kitimat Fjord System,
from Kitimat out to sea.
}
\keyword{datasets}

