% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/solve2lines.R
\name{solve2lines}
\alias{solve2lines}
\title{Find solution of 2 intersecting lines}
\usage{
solve2lines(line1, line2, toplot = TRUE)
}
\arguments{
\item{line1}{A list containing two coordinate pairs defining the first line segment. Format: \code{list(c(X1,Y1),c(X2,Y2))}.}

\item{line2}{A list containing two coordinate pairs defining the second line segment. Format: \code{list(c(X3,Y3),c(X4,Y4))}.}

\item{toplot}{If\code{TRUE} (the default), then a plot will be generated to show the lines and their intersection point. Plot range is defined by the end points of the lines.}
}
\value{
A 2-element vector, the X and Y of the intersection point. If there is no solution (i.e., parallel lines), both elements are \code{NA}.
}
\description{
Find the intersection point of two lines.
}
\details{
Developed for the function \code{\link{whalemap}} to test whether a bearing line from a vessel to a whale sighting intersects the shoreline before the horizon.
In that scenario, \code{line1} would be the bearing line between the vessel and the whale. \code{line2} would be formed between two points in a shoreline polygon.
This function would be looped to test for intersection between all shoreline points. It does not mess with Great-Sphere distances.
}
\examples{
line1 <- list(c(-129,53),c(-129.5,53.2))
line2 <- list(c(-128.5,53.12),c(-129.15,53.62))
solve2lines(line1,line2,toplot=TRUE)
}
\author{
Eric Keen, Scripps Institution of Oceanography, \email{ekeen@ucsd.edu}
}
\seealso{
\code{\link{whalemap}}
}

