% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-doc-subblocks.R
\docType{data}
\name{subblocks}
\alias{subblocks}
\title{Polygons of coordinates delineating 26 geographic subblocks in the Kitimat Fjord System}
\format{A data frame with 483,287 rows and 5 columns, formatted for seamless mapping with the PBSmapping package.}
\source{
Written and compiled by Eric Keen
}
\usage{
subblocks
}
\description{
A dataframe of coordinates for mapping 26 geographic blocks used to partition the Kitimat
Fjord System in Bangarang analyses.
}
\keyword{datasets}

