% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc-waterpolygons.R
\docType{data}
\name{waterpolygons}
\alias{waterpolygons}
\title{Polygons of shoreline coordinates delineating 54 geographic subblocks in the Kitimat Fjord System}
\format{
A data frame with 4458 rows and 5 columns:
}
\source{
Written and compiled by Eric Keen
}
\usage{
waterpolygons
}
\description{
A dataframe of coordinates for the mapping 54 geographic miniblocks used to partition the Kitimat 
Fjord System in Bangarang analyses.
}
\keyword{datasets}
