% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc-whale_sightings.R
\docType{data}
\name{whale_sightings}
\alias{whale_sightings}
\title{Metadata for 5km segments of Bangarang effort, 2013 - 2015.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 549 rows and 15 columns.
}
\usage{
whale_sightings
}
\description{
Metadata for 5km segments of Bangarang effort, 2013 - 2015.
}
\keyword{datasets}
