% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whalemap.R
\name{whalemap}
\alias{whalemap}
\title{Find true position of sighting}
\usage{
whalemap(X, Y, bearing, reticle, eye.height, vessel.hdg = NA, mag = -18.25,
  deg.per.ret = 0.279, toplot = FALSE)
}
\arguments{
\item{X}{Observer longitude (decimal degrees, where W is negative, e.g. -129.3 for the Kitimat Fjord System).}

\item{Y}{Observer latitude (decimal degrees).}

\item{bearing}{Magnetic bearing to sighting, in degrees. This is true bearing, not referenced to the vessel's heading.}

\item{reticle}{Reticle reading (must be positive). #'}

\item{eye.height}{Height above sea level of observer's eyes, in meters.}

\item{vessel.hdg}{The vessel's heading at the time of the sighting (Optional). It is assumed this information is
taken from a GPS, and therefore no magnetic declination correction is applied to it.
If provided, trackline distance to the sighting will be calculated, which assumes that the vessel is
headed along the trackline at the time of sighting.}

\item{mag}{Magnetic declination, in degrees, used to correct bearing readings from your magnetic compass.
Default is -18.25 deg, which is the approximate declination in the Kitimat Fjord System, British Columbia.}

\item{deg.per.ret}{Reticle-to-degree conversion. Default is 0.279,
which is specific to Fujinon 7x50 binocs (see references).
On the \emph{Bangarang}, we treated a full reticle as the distance between
long and short reticle lines in the optic of our Fujinon FMTRC-SX 7x50 binocs
(after Kinzey \& Gerrodette 2001).
This value needs to replaced with the correct conversion for the optics you use.}

\item{toplot}{When TRUE (not the default), a map will be generated that shows the process at work.}
}
\value{
Returns a list with the following elements:
\describe{
\item{\code{$X}}{X of animal}
\item{\code{$Y}}{Y of animal}
\item{\code{$radial.dist}}{Distance from observer to animal, in km (radial distance)}
\item{\code{$boundary}}{Reference boundary used (\code{"shore"} or \code{"horizon"})}
\item{\code{$boundary.dist}}{Observer's distance to boundary, in km}
\item{\code{$perp.dist}}{Trackline distance to sighting, in km (if vessel heading was provided). Ignores curvature of the earth.}
\item{\code{$track.bearing}}{Difference between sighting bearing and vessel heading, in degrees (simple subtraction)}
}
If \code{topot=TRUE} (not the DEFAULT), the function also generates a map key of the process. Key:
Black dot = Observer position;
Blue cross = Location of the horizon along the bearing line, based on eye height;
Blue line = Bearing line, from observer to the horizon based on eye height;
Green squares = (If any) Show nearby shoreline segments that were tested for obstruction of the bearing line;
Green lines = (If any) they show the shoreline segments that were tested for intersection with the bearing line before an intersection was found;
Dotted line = Vessel heading;
Red circles = Location of whale! Go cats!
}
\description{
Calculates the true position of a sighting in the northeast Pacific, in offshore or confined coastal waters.
}
\details{
This function can work offshore or within confined coastal zones, as tests to see whether the reticle
reading is referenced to the horizon or to shore. Also calculates the trackline distance of the
sighting, if vessel heading is provided, for use in distance sampling analysis.
}
\examples{
###### SHORE EXAMPLE ######
# You are on your way to Cameron Cove
# and see a whale in Taylor Bight:
whalemap(-129.21,53.0855,
        bearing=60,
        reticle=.1,
        eye.height=3.46,
        vessel.hdg=90,
        toplot=TRUE)

### HORIZON EXAMPLE ###############
# You are on your way back from the Wall
# and you see the bubble net group!
whalemap(X=-129.35577392578125,
         Y=52.860217383433515,
         bearing=14,
         reticle=.01,
         eye.height=3.46,
         vessel.hdg=45,
         toplot=TRUE)
}
\author{
Eric Keen, Scripps Institution of Oceanography, \email{ekeen@ucsd.edu}
}
\references{
\enumerate{
  \item Kinzey, D. and Gerrodette, T. 2001. Conversion factors for binocular reticles.
  Mar. Mammal Sci. 17(2):353-61.
  \item Lerczak, J.A. and Hobbs, R.C. 1998. Calculating sighting distances
from angular readings during shipboard, aerial, and shore-based marine mammal surveys.
Mar. Mammal Sci. 14(3):590-9. [See Errata. 1998. Mar. Mammal Sci. 14(4):903].
}
}
\seealso{
\code{\link[bangarang]{solve2lines}}
}

