% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc-segments_5km.R
\docType{data}
\name{segments_5km}
\alias{segments_5km}
\title{Metadata for 5km segments of Bangarang effort, 2013 - 2015.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 713 rows and 18 columns.
}
\usage{
segments_5km
}
\description{
Metadata for 5km segments of Bangarang effort, 2013 - 2015.
}
\keyword{datasets}
